chore: added pre-commit
