+++ .gitignore
@@ -3,4 +3,3 @@
 /tmp
 .env
 results.json
-
