+++ benches/benchmark_indices.rs
@@ -4,48 +4,40 @@ use rand::{
 };
 use rand_chacha::ChaCha8Rng;
 use serde::Serialize;
-use std::{
-    collections::HashMap,
-    env,
-    fs::File,
-    path::{
-        Path,
-        PathBuf,
-    },
-    time::{
-        Duration,
-        Instant,
-    },
+use std::collections::HashMap;
+use std::env;
+use std::fs::File;
+use std::path::{
+    Path,
+    PathBuf,
 };
-use timsquery::{
-    models::{
-        aggregators::RawPeakIntensityAggregator,
-        indices::{
-            expanded_raw_index::ExpandedRawFrameIndex,
-            raw_file_index::RawFileIndex,
-            transposed_quad_index::QuadSplittedTransposedIndex,
-        },
-    },
-    queriable_tims_data::queriable_tims_data::query_multi_group,
-    traits::tolerance::{
-        DefaultTolerance,
-        MobilityTolerance,
-        MzToleramce,
-        QuadTolerance,
-        RtTolerance,
-    },
-    ElutionGroup,
+use std::time::{
+    Duration,
+    Instant,
 };
+use timsquery::models::aggregators::RawPeakIntensityAggregator;
+use timsquery::models::indices::expanded_raw_index::ExpandedRawFrameIndex;
+use timsquery::models::indices::raw_file_index::RawFileIndex;
+use timsquery::models::indices::transposed_quad_index::QuadSplittedTransposedIndex;
+use timsquery::queriable_tims_data::queriable_tims_data::query_multi_group;
+use timsquery::traits::tolerance::{
+    DefaultTolerance,
+    MobilityTolerance,
+    MzToleramce,
+    QuadTolerance,
+    RtTolerance,
+};
+use timsquery::ElutionGroup;
 use tracing::subscriber::set_global_default;
 use tracing_bunyan_formatter::{
     BunyanFormattingLayer,
     JsonStorageLayer,
 };
 use tracing_chrome::ChromeLayerBuilder;
+use tracing_subscriber::prelude::*;
+use tracing_subscriber::registry::Registry;
 use tracing_subscriber::{
     fmt,
-    prelude::*,
-    registry::Registry,
     EnvFilter,
     Layer,
 };
