+++ rustfmt.toml
@@ -1,6 +1,6 @@
 
 imports_layout = "Vertical"
-imports_granularity = "Crate"
+imports_granularity = "Module"
 normalize_comments = true
 reorder_impl_items = true
 version = "Two"
