+++ src/lib.rs
@@ -1,17 +1,13 @@
 // Re-export main structures
-pub use crate::models::{
-    elution_group::ElutionGroup,
-    indices::transposed_quad_index::QuadSplittedTransposedIndex,
-};
+pub use crate::models::elution_group::ElutionGroup;
+pub use crate::models::indices::transposed_quad_index::QuadSplittedTransposedIndex;
 
 // Re-export traits
-pub use crate::traits::{
-    aggregator::Aggregator,
-    queriable_data::QueriableData,
-    tolerance::{
-        Tolerance,
-        ToleranceAdapter,
-    },
+pub use crate::traits::aggregator::Aggregator;
+pub use crate::traits::queriable_data::QueriableData;
+pub use crate::traits::tolerance::{
+    Tolerance,
+    ToleranceAdapter,
 };
 
 // Declare modules
