+++ src/models/indices/raw_file_index.rs
@@ -1,38 +1,27 @@
+use crate::models::adapters::FragmentIndexAdapter;
+use crate::models::frames::raw_frames::frame_elems_matching;
+use crate::models::frames::raw_peak::RawPeak;
+use crate::models::queries::{
+    FragmentGroupIndexQuery,
+    MsLevelContext,
+};
+use crate::traits::aggregator::Aggregator;
+use crate::traits::queriable_data::QueriableData;
 use crate::{
-    models::{
-        adapters::FragmentIndexAdapter,
-        frames::{
-            raw_frames::frame_elems_matching,
-            raw_peak::RawPeak,
-        },
-        queries::{
-            FragmentGroupIndexQuery,
-            MsLevelContext,
-        },
-    },
-    traits::{
-        aggregator::Aggregator,
-        queriable_data::QueriableData,
-    },
-    utils::tolerance_ranges::IncludedRange,
     ElutionGroup,
     ToleranceAdapter,
 };
 use rayon::iter::ParallelIterator;
 use serde::Serialize;
-use std::{
-    fmt::Debug,
-    hash::Hash,
+use std::fmt::Debug;
+use std::hash::Hash;
+use timsrust::readers::{
+    FrameReader,
+    FrameReaderError,
+    MetadataReader,
 };
 use timsrust::{
-    converters::ConvertableDomain,
-    readers::{
-        FrameReader,
-        FrameReaderError,
-        MetadataReader,
-    },
     Frame,
-    Metadata,
     TimsRustError,
 };
 use tracing::trace;
