+++ src/models/indices/transposed_quad_index/quad_splitted_transposed_index.rs
@@ -2,71 +2,56 @@ use super::quad_index::{
     TransposedQuadIndex,
     TransposedQuadIndexBuilder,
 };
-use crate::{
-    errors::Result,
-    models::{
-        adapters::FragmentIndexAdapter,
-        elution_group::ElutionGroup,
-        frames::{
-            expanded_frame::{
-                par_read_and_expand_frames,
-                ExpandedFrameSlice,
-                FrameProcessingConfig,
-                SortingStateTrait,
-            },
-            peak_in_quad::PeakInQuad,
-            raw_peak::RawPeak,
-            single_quad_settings::{
-                get_matching_quad_settings,
-                SingleQuadrupoleSetting,
-                SingleQuadrupoleSettingIndex,
-            },
-        },
-        queries::{
-            FragmentGroupIndexQuery,
-            MsLevelContext,
-        },
-    },
-    traits::{
-        aggregator::Aggregator,
-        queriable_data::QueriableData,
-    },
-    utils::{
-        display::{
-            glimpse_vec,
-            GlimpseConfig,
-        },
-        tolerance_ranges::IncludedRange,
-    },
-    ToleranceAdapter,
+use crate::errors::Result;
+use crate::models::adapters::FragmentIndexAdapter;
+use crate::models::elution_group::ElutionGroup;
+use crate::models::frames::expanded_frame::{
+    par_read_and_expand_frames,
+    ExpandedFrameSlice,
+    FrameProcessingConfig,
+    SortingStateTrait,
 };
+use crate::models::frames::peak_in_quad::PeakInQuad;
+use crate::models::frames::raw_peak::RawPeak;
+use crate::models::frames::single_quad_settings::{
+    get_matching_quad_settings,
+    SingleQuadrupoleSetting,
+    SingleQuadrupoleSettingIndex,
+};
+use crate::models::queries::{
+    FragmentGroupIndexQuery,
+    MsLevelContext,
+};
+use crate::traits::aggregator::Aggregator;
+use crate::traits::queriable_data::QueriableData;
+use crate::utils::display::{
+    glimpse_vec,
+    GlimpseConfig,
+};
+use crate::utils::tolerance_ranges::IncludedRange;
+use crate::ToleranceAdapter;
 use rayon::prelude::*;
 use serde::Serialize;
-use std::{
-    collections::HashMap,
-    fmt::{
-        Debug,
-        Display,
-    },
-    hash::Hash,
-    time::Instant,
+use std::collections::HashMap;
+use std::fmt::{
+    Debug,
+    Display,
 };
-use timsrust::{
-    converters::{
-        Scan2ImConverter,
-        Tof2MzConverter,
-    },
-    readers::{
-        FrameReader,
-        MetadataReader,
-    },
-    Metadata,
+use std::hash::Hash;
+use std::time::Instant;
+use timsrust::converters::{
+    Scan2ImConverter,
+    Tof2MzConverter,
 };
+use timsrust::readers::{
+    FrameReader,
+    MetadataReader,
+};
+use timsrust::Metadata;
 use tracing::{
     debug,
     info,
     instrument,
-    trace,
 };
 
 // TODO break this module apart ... its getting too big for my taste
@@ -287,7 +272,7 @@ impl QuadSplittedTransposedIndexBuilder {
         // TODO use the rayon contructor to fold
         let out2: Result<Vec<Self>> = split_frames
             .into_par_iter()
-            .map(|(q, frameslices)| {
+            .map(|(_q, frameslices)| {
                 // TODO:Refactor so the internal index is built first and then added.
                 // This should save a couple of thousand un-necessary hashmap lookups.
                 let mut out = Self::new();
